+++ src/UrlGenerator/UrlGenerator.php
@@ -5,6 +5,7 @@ namespace Innmind\Router\UrlGenerator;
 
 use Innmind\Router\{
     UrlGenerator as UrlGeneratorInterface,
+    Under,
     Route,
     Route\Name,
     Exception\NoMatchingRouteFound,
@@ -17,11 +18,11 @@ use Innmind\Immutable\{
 
 final class UrlGenerator implements UrlGeneratorInterface
 {
-    /** @var Sequence<Route> */
+    /** @var Sequence<Route|Under> */
     private Sequence $routes;
 
     /**
-     * @param Sequence<Route> $routes
+     * @param Sequence<Route|Under> $routes
      */
     public function __construct(Sequence $routes)
     {
@@ -32,6 +33,10 @@ final class UrlGenerator implements UrlGeneratorInterface
     {
         return $this
             ->routes
+            ->flatMap(static fn($route) => match (true) {
+                $route instanceof Under => $route->routes(),
+                default => Sequence::of($route),
+            })
             ->find(static fn(Route $candidate): bool => $candidate->is($route))
             ->map(static fn($route) => $route->template())
             ->map(static fn($template) => $template->expand($variables ?? Map::of()))