+++ tests/RequestMatcher/RequestMatcherTest.php
@@ -8,6 +8,7 @@ use Innmind\Router\{
     RequestMatcher as RequestMatcherInterface,
     Route,
     Route\Name,
+    Under,
 };
 use Innmind\Http\{
     ServerRequest,
@@ -17,7 +18,7 @@ use Innmind\Http\{
 use Innmind\UrlTemplate\Template;
 use Innmind\Url\Url;
 use Innmind\Immutable\Sequence;
-use PHPUnit\Framework\TestCase;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class RequestMatcherTest extends TestCase
 {
@@ -69,5 +70,65 @@ class RequestMatcherTest extends TestCase
             static fn($route) => $route,
             static fn() => null,
         ));
+    }
+
+    public function testMatchGroupedRoutes()
+    {
+        $match = new RequestMatcher(
+            Sequence::of(
+                Under::of(Template::of('/foo'))
+                    ->route(Method::delete)
+                    ->route(Method::post, static fn($route) => $route->named(Name::of('foo')))
+                    ->route(Method::get),
+                Under::of(Template::of('/bar'))
+                    ->route(Method::delete)
+                    ->route(Method::post, static fn($route) => $route->named(Name::of('bar')))
+                    ->route(Method::get),
+            ),
+        );
+        $foo = ServerRequest::of(
+            Url::of('/foo'),
+            Method::post,
+            ProtocolVersion::v11,
+        );
+        $bar = ServerRequest::of(
+            Url::of('/bar'),
+            Method::post,
+            ProtocolVersion::v11,
+        );
+
+        $this->assertTrue($match($foo)->match(
+            static fn($route) => $route->is(Name::of('foo')),
+            static fn() => null,
+        ));
+        $this->assertTrue($match($bar)->match(
+            static fn($route) => $route->is(Name::of('bar')),
+            static fn() => null,
+        ));
+    }
+
+    public function testMatchNotAllowedMethod()
+    {
+        $match = new RequestMatcher(
+            Sequence::of(
+                Under::of(Template::of('/foo'))
+                    ->route(Method::delete)
+                    ->route(Method::post)
+                    ->route(Method::get),
+            ),
+        );
+        $request = ServerRequest::of(
+            Url::of('/foo'),
+            Method::head,
+            ProtocolVersion::v11,
+        );
+
+        $response = $match($request)->match(
+            static fn($route) => $route->respondTo($request),
+            static fn() => null,
+        );
+
+        $this->assertNotNull($response);
+        $this->assertSame(405, $response->statusCode()->toInt());
     }
 }